fix: drill into treemap folders at arbitrary depth

+++ crates/voidspace-app/src/treemap_state.rs
@@ -3,7 +3,7 @@ use voidspace_model::NodeId;
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct AggregateSelection {
     pub parent: NodeId,
-    pub depth: u8,
+    pub depth: usize,
     pub members: Vec<NodeId>,
 }
